Added sale_price functionality

+++ website_app/models.py
@@ -10,7 +10,7 @@ class product(models.Model):
     description = models.TextField(null=True, blank=True)
     image = models.ImageField(upload_to='product_images/', null=True, blank=True)
     price = models.IntegerField(null=True, blank=True)
-    sale_price = models.FloatField(null=True, blank=True)
+    sale_price = models.IntegerField(null=True, blank=True)
     material = models.CharField(max_length=200, null=True, blank=True)
     color = models.CharField(max_length=200, null=True, blank=True)
     height = models.CharField(max_length=200, null=True, blank=True)
